Basic code throwdown for how we update weights
Also using proper message types. I realize we're gonna want to fake our
feature vectors for testing, so let's sub to a topic and do it that way.
in other words, in the features used to calculate cost (not update
weights, that's different) will come in from some topic. haven't
implemented it yet but that's how i'll do it