Fixed a bunch of bugs and added bias
The idea is that each bucket for each feature has both a weight and a
bias. The cost is now feature_value * weight + bias. This means we can
learn the cost of buckets where the feature_value is 0, because
otherwise we would just always get 0 weight update. We update bias the
same way we update weight. The difference is that weight is
feature_value*learning rate, which is zero if feature_value is 0. But
our bias always has a feature_value of 1, so it's always updated by
exactly the learning_rate.

Still haven't implemented any faking of state input. Weights are
updated, but you can't calculate cost without an feature vector for
your current state.