Snapshotting because it worked!
I showed the the cost of some odom path would increase for areas along
demonstrations when the robot state matches the state in which the
demonstration occured. In other words, I draw where the robot went
(as if mocking odom data) and my demo. I then say all that occured when
the goal was some number. Weights for the bucket containing that goal
are updated. When the robot is in that state (same goal is set), then
the cost will reflet the demo. When the robot is in another state (goal
in another bucket) they will not. I showed that multiple demos could be
given in different states, and weights could be individual to each
macrocell/state combination.

Lots of more testing to do, and also I haven't solved the issue of
decreaseing cost