Implemented new normalizing scheme
to preserve relative changes in weight (a lot vs a little change) we
don't normalize between 0 and 128, we just add the min_cost to
everything. This means after the first update (lets pretend it updated
the min costs to -2 and 2), you'd have cost ranging from 0 to 4. There
is a hard cap at 128.